--- Ex3/README.md
@@ -1,11 +1,0 @@
-# Ariel_OOP_2020
-The course is designed as a "project-paced-learning" (aka BPL) it consists of
-5-6 Object Oriented Designing and Programming assignments - which will teach (hands on) the students
-good practice of OOP, OOD, Agile (Xtream) Programming, while working on large scale software project
-Using Github.         
-
-
-
-
-
-
