--- Ex3/data/A4
@@ -1,1 +1,0 @@
-{"Edges":[{"src":0,"w":1.0835435956906543,"dest":1},{"src":0,"w":0.3433607693232187,"dest":22},{"src":1,"w":1.260431259778731,"dest":0},{"src":1,"w":1.0123064430949418,"dest":2},{"src":2,"w":1.7247835531480924,"dest":1},{"src":2,"w":1.9578538886727144,"dest":3},{"src":2,"w":1.3013251948418596,"dest":30},{"src":3,"w":1.827752726467304,"dest":2},{"src":3,"w":1.9908441205772092,"dest":4},{"src":4,"w":1.6284986709004998,"dest":3},{"src":4,"w":1.1712767410444496,"dest":5},{"src":5,"w":1.0131908397905627,"dest":4},{"src":5,"w":1.9446655683736058,"dest":6},{"src":5,"w":1.3874752630258282,"dest":25},{"src":6,"w":1.560477950671761,"dest":5},{"src":6,"w":1.5464614947923163,"dest":7},{"src":6,"w":1.134232756788375,"dest":15},{"src":7,"w":1.4794484033472592,"dest":32},{"src":7,"w":1.9337075379943074,"dest":6},{"src":7,"w":1.7158358413722645,"dest":8},{"src":8,"w":1.130370301760947,"dest":35},{"src":8,"w":1.228339449624865,"dest":7},{"src":8,"w":1.2827103384274627,"dest":9},{"src":9,"w":1.2340510488941445,"dest":8},{"src":9,"w":1.3573438076507416,"dest":10},{"src":10,"w":0.5657504395294557,"dest":39},{"src":10,"w":1.867530854380139,"dest":9},{"src":10,"w":1.7841169947193596,"dest":11},{"src":11,"w":1.8878405720440825,"dest":10},{"src":11,"w":1.6627775783718608,"dest":12},{"src":12,"w":1.668568602308967,"dest":11},{"src":12,"w":1.3618252445279904,"dest":13},{"src":13,"w":1.2199921006106256,"dest":12},{"src":13,"w":1.8985651437927196,"dest":14},{"src":14,"w":1.30672526242022,"dest":17},{"src":14,"w":1.7334336008841222,"dest":13},{"src":14,"w":1.996989332206076,"dest":15},{"src":15,"w":1.8312618692355023,"dest":16},{"src":15,"w":1.32134381322386,"dest":6},{"src":15,"w":1.350338856626182,"dest":14},{"src":16,"w":1.275567351461944,"dest":17},{"src":16,"w":1.6032568850637063,"dest":19},{"src":16,"w":1.0913562448992056,"dest":15},{"src":17,"w":1.380583524414519,"dest":16},{"src":17,"w":1.895783214212016,"dest":18},{"src":17,"w":1.280709814240381,"dest":14},{"src":18,"w":1.3638435750384605,"dest":17},{"src":18,"w":1.5927402586330097,"dest":19},{"src":19,"w":1.2406524918828385,"dest":16},{"src":19,"w":1.809592242888764,"dest":18},{"src":19,"w":1.763756279531104,"dest":20},{"src":20,"w":1.4313420158759202,"dest":19},{"src":20,"w":1.1472916434588263,"dest":21},{"src":21,"w":1.4108678669726202,"dest":20},{"src":21,"w":1.8457046682315756,"dest":22},{"src":22,"w":0.44931229584749066,"dest":0},{"src":22,"w":1.7149845299642799,"dest":21},{"src":22,"w":1.0956665867713435,"dest":23},{"src":23,"w":1.20813159861614,"dest":22},{"src":23,"w":1.9244056090500044,"dest":24},{"src":24,"w":1.366805158502689,"dest":23},{"src":24,"w":1.9409302255351337,"dest":25},{"src":25,"w":1.5302621186016696,"dest":5},{"src":25,"w":1.526540220474201,"dest":24},{"src":25,"w":1.8867184054487574,"dest":26},{"src":26,"w":1.886708122591454,"dest":25},{"src":26,"w":1.5971039511349687,"dest":27},{"src":27,"w":1.814170129663763,"dest":26},{"src":27,"w":1.5757658838715223,"dest":28},{"src":28,"w":1.54852926169781,"dest":27},{"src":28,"w":1.5978560015720404,"dest":29},{"src":28,"w":1.531973599491279,"dest":30},{"src":29,"w":1.3870892671587354,"dest":28},{"src":29,"w":1.320315386625936,"dest":30},{"src":30,"w":1.4857584689137417,"dest":2},{"src":30,"w":1.0201144759650802,"dest":28},{"src":30,"w":1.1134294319752716,"dest":29},{"src":30,"w":1.5505858880979928,"dest":31},{"src":31,"w":1.3935717941599135,"dest":32},{"src":31,"w":1.961523838549256,"dest":30},{"src":32,"w":1.3524096594998078,"dest":33},{"src":32,"w":1.1072198955013588,"dest":7},{"src":32,"w":1.9566531707928996,"dest":31},{"src":33,"w":1.633561638256882,"dest":32},{"src":33,"w":1.8111759305323711,"dest":34},{"src":33,"w":1.097054390982287,"dest":35},{"src":34,"w":1.304814668643914,"dest":33},{"src":34,"w":1.665073852199095,"dest":35},{"src":35,"w":1.1355123315390185,"dest":33},{"src":35,"w":1.8300333272322125,"dest":34},{"src":35,"w":1.8365475807719966,"dest":36},{"src":35,"w":1.180655161883633,"dest":8},{"src":36,"w":1.9573831475558512,"dest":35},{"src":36,"w":1.1918019631569365,"dest":37},{"src":37,"w":1.1439761461732751,"dest":36},{"src":37,"w":1.1004358793841362,"dest":38},{"src":37,"w":0.8728663213495164,"dest":39},{"src":38,"w":1.493642552290904,"dest":37},{"src":38,"w":1.6111110787665157,"dest":39},{"src":39,"w":0.9832912320283205,"dest":37},{"src":39,"w":1.6032641568791934,"dest":38},{"src":39,"w":0.7421323463930005,"dest":10}],"Nodes":[{"pos":"35.18750930912026,32.10374591260504,0.0","id":0},{"pos":"35.18961076190476,32.10794390084033,0.0","id":1},{"pos":"35.19338913155771,32.10616293613445,0.0","id":2},{"pos":"35.19742222276029,32.1051815882353,0.0","id":3},{"pos":"35.1988444180791,32.103727739495795,0.0","id":4},{"pos":"35.19589389346247,32.1015651394958,0.0","id":5},{"pos":"35.20332327199354,32.103255238655464,0.0","id":6},{"pos":"35.2016888087167,32.10592668571429,0.0","id":7},{"pos":"35.20574312671509,32.10625380168067,0.0","id":8},{"pos":"35.20797194027441,32.10481812605042,0.0","id":9},{"pos":"35.21219607263922,32.10616293613445,0.0","id":10},{"pos":"35.21168662953995,32.10498168403361,0.0","id":11},{"pos":"35.21315127845036,32.10427293277311,0.0","id":12},{"pos":"35.21032811460856,32.10112898487395,0.0","id":13},{"pos":"35.20765353833737,32.10245562184874,0.0","id":14},{"pos":"35.204809147699756,32.103073507563025,0.0","id":15},{"pos":"35.20459687974173,32.10154696638656,0.0","id":16},{"pos":"35.20737758999193,32.101092638655466,0.0","id":17},{"pos":"35.20616766263115,32.09920263529412,0.0","id":18},{"pos":"35.204257251008876,32.09980234789916,0.0","id":19},{"pos":"35.199353861178366,32.100474752941174,0.0","id":20},{"pos":"35.19084191606134,32.10134706218487,0.0","id":21},{"pos":"35.18787016464891,32.10347331596639,0.0","id":22},{"pos":"35.18950462792575,32.10519976134454,0.0","id":23},{"pos":"35.191924482647295,32.104200240336134,0.0","id":24},{"pos":"35.195575491525425,32.10349148907563,0.0","id":25},{"pos":"35.18806120581114,32.099711482352944,0.0","id":26},{"pos":"35.18725458757062,32.109306884033614,0.0","id":27},{"pos":"35.19483255367231,32.10943409579832,0.0","id":28},{"pos":"35.19243392574657,32.10854361344538,0.0","id":29},{"pos":"35.194195749798226,32.10774399663865,0.0","id":30},{"pos":"35.199523675544796,32.10908880672269,0.0","id":31},{"pos":"35.20115813882163,32.10750774621849,0.0","id":32},{"pos":"35.20476669410815,32.108579959663864,0.0","id":33},{"pos":"35.20769599192897,32.107816689075634,0.0","id":34},{"pos":"35.206358703793384,32.10739870756303,0.0","id":35},{"pos":"35.2099248054883,32.10759861176471,0.0","id":36},{"pos":"35.21119841323648,32.10794390084033,0.0","id":37},{"pos":"35.21217484584342,32.10854361344538,0.0","id":38},{"pos":"35.212620608555284,32.10719880336134,0.0","id":39}]}
