--- Ex3/data/G_10_80_0.json
@@ -1,1 +1,0 @@
-{"Edges":[{"src":0,"w":45.970607092631596,"dest":0},{"src":0,"w":78.71917176578098,"dest":2},{"src":0,"w":54.25386036035215,"dest":3},{"src":0,"w":38.26679232313723,"dest":5},{"src":0,"w":35.29781381114986,"dest":6},{"src":0,"w":72.29045089185382,"dest":8},{"src":0,"w":76.7849885135893,"dest":9},{"src":1,"w":60.85181760699909,"dest":0},{"src":1,"w":54.370932860912006,"dest":2},{"src":1,"w":45.66512032063395,"dest":3},{"src":1,"w":44.36834230854706,"dest":4},{"src":1,"w":60.01099833651171,"dest":5},{"src":1,"w":3.6258840874618037,"dest":6},{"src":1,"w":34.16023404618369,"dest":7},{"src":1,"w":48.121077297163005,"dest":8},{"src":1,"w":12.991389377500887,"dest":9},{"src":2,"w":60.17070392240527,"dest":0},{"src":2,"w":63.22224597685683,"dest":1},{"src":2,"w":75.22482775424128,"dest":5},{"src":2,"w":50.792364662093604,"dest":6},{"src":2,"w":46.93853965870328,"dest":8},{"src":2,"w":19.295692442460734,"dest":9},{"src":3,"w":33.15580601566146,"dest":0},{"src":3,"w":16.693851970959226,"dest":2},{"src":3,"w":70.1712210669341,"dest":3},{"src":3,"w":12.269239945821651,"dest":5},{"src":3,"w":49.73668882050413,"dest":7},{"src":3,"w":9.161792026701495,"dest":8},{"src":3,"w":37.69777463191115,"dest":9},{"src":4,"w":60.97620356286488,"dest":0},{"src":4,"w":68.4344026293197,"dest":1},{"src":4,"w":66.61581417150457,"dest":2},{"src":4,"w":9.878965843993443,"dest":3},{"src":4,"w":15.72319842169162,"dest":4},{"src":4,"w":64.29964250374522,"dest":5},{"src":4,"w":40.679369363356265,"dest":6},{"src":4,"w":62.443108718953006,"dest":7},{"src":4,"w":30.025587223630307,"dest":8},{"src":4,"w":25.31198039868959,"dest":9},{"src":5,"w":78.39099841740129,"dest":0},{"src":5,"w":23.034518005083623,"dest":1},{"src":5,"w":49.855467972050455,"dest":2},{"src":5,"w":10.383183147587454,"dest":3},{"src":5,"w":38.73527315771003,"dest":5},{"src":5,"w":1.804322983721461,"dest":6},{"src":5,"w":8.545509779174274,"dest":7},{"src":5,"w":22.42532344439192,"dest":8},{"src":5,"w":25.391773110322667,"dest":9},{"src":6,"w":18.00730774236018,"dest":0},{"src":6,"w":11.7221365212227,"dest":1},{"src":6,"w":33.97166314824226,"dest":2},{"src":6,"w":63.44454209273474,"dest":3},{"src":6,"w":44.85144195445204,"dest":5},{"src":6,"w":63.136947614556114,"dest":7},{"src":6,"w":47.95113564587882,"dest":8},{"src":6,"w":20.817019611467508,"dest":9},{"src":7,"w":8.053739541703226,"dest":2},{"src":7,"w":56.78599625217979,"dest":3},{"src":7,"w":13.675470857615256,"dest":4},{"src":7,"w":55.465892242421944,"dest":5},{"src":7,"w":36.85935451061333,"dest":6},{"src":7,"w":27.351069575193236,"dest":7},{"src":7,"w":71.57754201874826,"dest":8},{"src":7,"w":77.58467075634455,"dest":9},{"src":8,"w":71.05836712098343,"dest":0},{"src":8,"w":44.98277141335169,"dest":1},{"src":8,"w":64.25056334121751,"dest":2},{"src":8,"w":24.648277376911594,"dest":4},{"src":8,"w":0.9809245806217883,"dest":5},{"src":8,"w":69.43544191570957,"dest":6},{"src":8,"w":12.822367209039761,"dest":7},{"src":8,"w":52.95596902230983,"dest":8},{"src":8,"w":62.7223484749542,"dest":9},{"src":9,"w":74.40756016240704,"dest":2},{"src":9,"w":46.639978946640746,"dest":3},{"src":9,"w":48.843505594430404,"dest":4},{"src":9,"w":20.83803747728078,"dest":5},{"src":9,"w":35.94421527054947,"dest":6},{"src":9,"w":69.22256449325289,"dest":7},{"src":9,"w":71.57571659415565,"dest":9}],"Nodes":[{"id":0},{"id":1},{"id":2},{"id":3},{"id":4},{"id":5},{"id":6},{"id":7},{"id":8},{"id":9}]}
