--- Ex3/data/G_10_80_1.json
@@ -1,1 +1,0 @@
-{"Edges":[{"src":0,"w":79.10210421546576,"dest":0},{"src":0,"w":30.41763241754926,"dest":1},{"src":0,"w":5.222720639544136,"dest":2},{"src":0,"w":61.540474801988395,"dest":3},{"src":0,"w":1.4254324246464245,"dest":4},{"src":0,"w":29.62317157192879,"dest":6},{"src":0,"w":77.69682104395123,"dest":8},{"src":1,"w":59.768331257987676,"dest":1},{"src":1,"w":55.634319236347004,"dest":2},{"src":1,"w":46.09408949974762,"dest":3},{"src":1,"w":21.506152804579607,"dest":4},{"src":1,"w":32.542031972906635,"dest":6},{"src":1,"w":40.0879200213138,"dest":7},{"src":1,"w":78.75421229608374,"dest":9},{"src":2,"w":20.32927659218075,"dest":0},{"src":2,"w":33.09958724559113,"dest":1},{"src":2,"w":77.36171217493904,"dest":2},{"src":2,"w":1.3859528226530131,"dest":3},{"src":2,"w":19.148756160560012,"dest":5},{"src":2,"w":68.5725612585445,"dest":6},{"src":2,"w":34.03119476857865,"dest":7},{"src":2,"w":45.38199711844628,"dest":9},{"src":3,"w":79.44500875000745,"dest":0},{"src":3,"w":0.19844075676367479,"dest":2},{"src":3,"w":36.07861155423006,"dest":3},{"src":3,"w":11.051960610655769,"dest":4},{"src":3,"w":12.616628926759029,"dest":5},{"src":3,"w":25.290332627359888,"dest":6},{"src":3,"w":20.770897308737464,"dest":7},{"src":3,"w":43.22654923244013,"dest":8},{"src":3,"w":36.47021817121635,"dest":9},{"src":4,"w":44.630151839306954,"dest":0},{"src":4,"w":75.97510297236161,"dest":1},{"src":4,"w":54.360674481191516,"dest":2},{"src":4,"w":40.06550642488482,"dest":3},{"src":4,"w":14.208975717967066,"dest":4},{"src":4,"w":30.30431054074055,"dest":6},{"src":4,"w":65.67025086708223,"dest":7},{"src":4,"w":46.21390777683535,"dest":9},{"src":5,"w":10.330559391069462,"dest":0},{"src":5,"w":64.96788831783256,"dest":1},{"src":5,"w":1.2532893145708224,"dest":3},{"src":5,"w":44.63963719095241,"dest":4},{"src":5,"w":7.108400057476756,"dest":5},{"src":5,"w":70.62368856451643,"dest":6},{"src":5,"w":17.670240063160367,"dest":7},{"src":5,"w":33.54262532947976,"dest":8},{"src":6,"w":79.40127823775842,"dest":0},{"src":6,"w":20.114713803486012,"dest":2},{"src":6,"w":48.543943880737324,"dest":3},{"src":6,"w":51.93954623081834,"dest":4},{"src":6,"w":28.09251479413419,"dest":6},{"src":6,"w":40.169197318471696,"dest":7},{"src":6,"w":31.45128269162786,"dest":8},{"src":6,"w":42.42967001833669,"dest":9},{"src":7,"w":13.540448611094906,"dest":0},{"src":7,"w":70.446231081275,"dest":2},{"src":7,"w":61.39605204940575,"dest":4},{"src":7,"w":2.680329249746613,"dest":5},{"src":7,"w":11.518508280697107,"dest":6},{"src":7,"w":42.074654794490044,"dest":7},{"src":7,"w":8.571623994115237,"dest":8},{"src":7,"w":21.043957151829343,"dest":9},{"src":8,"w":36.21817515893986,"dest":0},{"src":8,"w":27.95723796652554,"dest":1},{"src":8,"w":64.68232930755588,"dest":3},{"src":8,"w":30.97565078156314,"dest":4},{"src":8,"w":18.735311643115136,"dest":5},{"src":8,"w":62.8711699606534,"dest":6},{"src":8,"w":71.5334551162375,"dest":7},{"src":8,"w":52.898835614304225,"dest":8},{"src":8,"w":80.03878406006915,"dest":9},{"src":9,"w":18.795560638889896,"dest":0},{"src":9,"w":52.16829691396408,"dest":1},{"src":9,"w":61.231755622547205,"dest":3},{"src":9,"w":52.32276569469346,"dest":4},{"src":9,"w":54.70505267245872,"dest":6},{"src":9,"w":8.796334140129307,"dest":7},{"src":9,"w":30.209681940090736,"dest":8},{"src":9,"w":24.353457934591034,"dest":9}],"Nodes":[{"pos":"1.0,0.0,0.0","id":0},{"pos":"0.8090169943749475,0.5877852522924731,0.0","id":1},{"pos":"0.30901699437494745,0.9510565162951535,0.0","id":2},{"pos":"-0.30901699437494734,0.9510565162951536,0.0","id":3},{"pos":"-0.8090169943749473,0.5877852522924732,0.0","id":4},{"pos":"-1.0,1.2246467991473532E-16,0.0","id":5},{"pos":"-0.8090169943749476,-0.587785252292473,0.0","id":6},{"pos":"-0.30901699437494756,-0.9510565162951535,0.0","id":7},{"pos":"0.30901699437494723,-0.9510565162951536,0.0","id":8},{"pos":"0.8090169943749473,-0.5877852522924734,0.0","id":9}]}
