--- Ex3/data/T0.json_saved
@@ -1,69 +1,0 @@
-{
-    "graph": {
-        "0": {
-            "key": 0,
-            "tag": -1,
-            "value": Infinity,
-            "pos": [
-                54.54768918357116,
-                80.25255464032239
-            ],
-            "inside": {
-                "1": 1.1
-            },
-            "outside": {
-                "1": 1
-            },
-            "prev": null
-        },
-        "1": {
-            "key": 1,
-            "tag": -1,
-            "value": Infinity,
-            "pos": [
-                4.721397716745301,
-                96.9850264065137
-            ],
-            "inside": {
-                "0": 1
-            },
-            "outside": {
-                "0": 1.1,
-                "2": 1.3,
-                "3": 1.8
-            },
-            "prev": null
-        },
-        "2": {
-            "key": 2,
-            "tag": -1,
-            "value": Infinity,
-            "pos": [
-                38.0957380944294,
-                95.04905946149489
-            ],
-            "inside": {
-                "1": 1.3
-            },
-            "outside": {
-                "3": 1.1
-            },
-            "prev": null
-        },
-        "3": {
-            "key": 3,
-            "tag": 1,
-            "value": 0,
-            "pos": [
-                2.1647538664965826,
-                86.95163472590149
-            ],
-            "inside": {
-                "1": 1.8,
-                "2": 1.1
-            },
-            "outside": {},
-            "prev": null
-        }
-    }
-}
