+++ data/G_10_80_2.json
@@ -1,0 +1,1 @@
+{"Edges":[{"src":0,"w":15.240153972118284,"dest":0},{"src":0,"w":25.068566978163332,"dest":2},{"src":0,"w":42.21788539643512,"dest":3},{"src":0,"w":55.58843919899137,"dest":4},{"src":0,"w":55.425375330029794,"dest":5},{"src":0,"w":37.01089239197957,"dest":6},{"src":0,"w":77.21057152591118,"dest":7},{"src":0,"w":38.87352979733993,"dest":8},{"src":0,"w":34.25892844218011,"dest":9},{"src":1,"w":70.93102978445458,"dest":0},{"src":1,"w":61.50378944986711,"dest":2},{"src":1,"w":20.18479087357172,"dest":3},{"src":1,"w":52.25971115311266,"dest":4},{"src":1,"w":68.61811266260328,"dest":5},{"src":1,"w":49.63906138499115,"dest":7},{"src":1,"w":50.33915463562098,"dest":8},{"src":1,"w":73.99313088662157,"dest":9},{"src":2,"w":25.266279540144282,"dest":0},{"src":2,"w":72.93625753958574,"dest":1},{"src":2,"w":66.74133788742238,"dest":2},{"src":2,"w":67.93689083773099,"dest":3},{"src":2,"w":52.824568033962414,"dest":4},{"src":2,"w":57.9373150393028,"dest":5},{"src":2,"w":18.553934040660664,"dest":8},{"src":2,"w":43.57336304008557,"dest":9},{"src":3,"w":75.32588232115813,"dest":0},{"src":3,"w":30.00066117774944,"dest":1},{"src":3,"w":55.989547660187434,"dest":2},{"src":3,"w":1.6984726493778244,"dest":3},{"src":3,"w":40.54373218317974,"dest":4},{"src":3,"w":29.041801466953938,"dest":5},{"src":3,"w":79.75655435572536,"dest":7},{"src":3,"w":14.19617663424928,"dest":8},{"src":3,"w":4.981550897410025,"dest":9},{"src":4,"w":46.742004361108926,"dest":0},{"src":4,"w":69.78290268787323,"dest":1},{"src":4,"w":76.73665975786902,"dest":2},{"src":4,"w":1.0675761333846803,"dest":3},{"src":4,"w":48.853959907900986,"dest":5},{"src":4,"w":58.25854655828068,"dest":6},{"src":4,"w":4.979443052452949,"dest":8},{"src":4,"w":20.30162823443138,"dest":9},{"src":5,"w":45.197697384217015,"dest":0},{"src":5,"w":1.1921334747461307,"dest":2},{"src":5,"w":10.26931581586253,"dest":3},{"src":5,"w":66.86553932817891,"dest":4},{"src":5,"w":51.93531659580427,"dest":8},{"src":6,"w":59.29291165111422,"dest":0},{"src":6,"w":66.32587721016867,"dest":2},{"src":6,"w":78.93442804967405,"dest":3},{"src":6,"w":21.298446435899784,"dest":4},{"src":6,"w":57.523147057397594,"dest":5},{"src":6,"w":13.265700654119568,"dest":7},{"src":6,"w":58.72831442711382,"dest":8},{"src":6,"w":72.33145032137395,"dest":9},{"src":7,"w":8.585198776841883,"dest":0},{"src":7,"w":36.7462516991881,"dest":1},{"src":7,"w":64.49293848561105,"dest":2},{"src":7,"w":8.301358102186478,"dest":3},{"src":7,"w":20.276390687061422,"dest":4},{"src":7,"w":37.415219605587986,"dest":5},{"src":7,"w":21.032992334148854,"dest":6},{"src":7,"w":33.3437067665264,"dest":7},{"src":7,"w":53.8936816407964,"dest":9},{"src":8,"w":29.688494483003844,"dest":0},{"src":8,"w":27.501571753368708,"dest":1},{"src":8,"w":17.469236118831454,"dest":2},{"src":8,"w":47.18600142245294,"dest":3},{"src":8,"w":71.42875859594682,"dest":4},{"src":8,"w":41.55017201608854,"dest":5},{"src":8,"w":39.65031695482115,"dest":6},{"src":8,"w":61.59502029781805,"dest":7},{"src":8,"w":44.45754240048374,"dest":8},{"src":8,"w":71.49519467846294,"dest":9},{"src":9,"w":4.027440982054508,"dest":0},{"src":9,"w":59.71060469479066,"dest":1},{"src":9,"w":24.92585800566622,"dest":3},{"src":9,"w":11.353131647629,"dest":4},{"src":9,"w":16.798636606017865,"dest":5},{"src":9,"w":23.689398380018293,"dest":6}],"Nodes":[{"pos":"0.08170009195216499,0.7541519252293025,0.0","id":0},{"pos":"0.8915613619809635,0.19891432952823085,0.0","id":1},{"pos":"0.002006127598024876,0.23308551122875265,0.0","id":2},{"pos":"0.2724319354710991,0.3490892016035897,0.0","id":3},{"pos":"0.0873338855096586,0.5200615085512579,0.0","id":4},{"pos":"0.9543604439583947,0.9625016485900687,0.0","id":5},{"pos":"0.021340414514527617,0.5499564766081825,0.0","id":6},{"pos":"0.9420675020620758,0.7598820416652761,0.0","id":7},{"pos":"0.5026280640578298,0.06167767495113918,0.0","id":8},{"pos":"0.3161793602001721,0.818961443942509,0.0","id":9}]}
