+++ data/T0.json
@@ -1,0 +1,1 @@
+{"Nodes": [{"id": 0}, {"id": 1}, {"id": 2}, {"id": 3}], "Edges": [{"src": 0, "dest": 1, "w": 1}, {"src": 1, "dest": 0, "w": 1.1}, {"src": 1, "dest": 2, "w": 1.3}, {"src": 1, "dest": 3, "w": 1.8}, {"src": 2, "dest": 3, "w": 1.1}]}
